chore: rename store folder as stores
