+++ web/app/components/Organism/NavigationDrawer/index.vue
@@ -1,7 +1,6 @@
 <!-- SPDX-License-Identifier: AGPL-3.0-or-later -->
 <script setup lang="ts">
 import type { NavigationDrawerListProps } from './List/index.vue'
-import { useLayoutStore } from '~/store/layout.store'
 
 defineProps<{
   logoutLoading: boolean
